Users' rights

+++ Trello/src/app/boards.component/boards.component.ts
@@ -6,6 +6,7 @@ import { BoardsService } from '../services/boards.service';
 import { BoardService } from '../services/board.service';
 import { UsersService } from '../services/users.service';
 import { JwtHelper } from 'angular2-jwt';
+import { Router } from '@angular/router';
 
 @Component({
     selector: 'boards',
@@ -18,20 +19,19 @@ export class BoardsComponent {
     newBoardName: string;
     jwtHelper: JwtHelper = new JwtHelper();
     tokens: any = this.authService.getTokens();
+    rights: string;
     constructor(
         private backendService: BackendService,
         private authService: AuthService,
         private boardsService: BoardsService,
         private usersService: UsersService,
-        private boardService: BoardService
+        private boardService: BoardService,
+        private router: Router
     ) { }
     addBoard(): void {
-        if (!this.newBoardName) { return; }
-        // this.boards.push(new Board(+new Date(), this.newBoardName, []));        
+        if (!this.newBoardName) { return; }     
         this.boardsService.addBoard(this.newBoardName).subscribe(
             data => {
-                // console.log('from add board');
-                // console.log(data);
                 this.usersService.setRights(this.jwtHelper.decodeToken(this.tokens.accessToken).id, data._id, 'owner').subscribe(
                     d => {console.log(d);}
                 );
@@ -45,9 +45,17 @@ export class BoardsComponent {
             });
     }
     removeBoard(board): void {
-        this.boards.splice(this.boards.findIndex((element) => element == board), 1);
-        this.boardService.deleteBoard(board.id).subscribe(
-            data => { console.log(data); });
+        this.usersService.getRights(UsersService.user.id, board.id).subscribe(
+            rights => {
+                console.log('got rights='+rights.rights);
+                if (rights.rights!='owner') {
+                   alert('No access rights!');
+                   return;
+                }
+                this.boards.splice(this.boards.findIndex((element) => element == board), 1);
+                this.boardService.deleteBoard(board.id).subscribe(
+                    data => { console.log(data); });
+                }); 
     }
     ngOnInit() {
         this.boardsService.getBoardsFromServer().subscribe(
@@ -56,7 +64,6 @@ export class BoardsComponent {
                 this.boards = this.boardsService.getBoards();
             },
             err => {
-                // console.log(err);
                 this.authService.refreshTokens(this.tokens.refreshToken).subscribe(
                     data => {
                         this.authService.setTokens(data);
@@ -67,6 +74,16 @@ export class BoardsComponent {
                             });
                     }
                 );
-            });
+            });        
+    }
+    chooseBoard(id): void {
+        this.usersService.getRights(UsersService.user.id, id).subscribe(
+            rights => {
+                console.log('got rights='+rights.rights);
+                if (rights.rights!='none') {
+                   this.router.navigate([`/board/${id}`]);
+                } else { alert('No access rights!'); }
+            }
+        );        
     }
 }
